fixed location of api call and imports
